refactor: try migrating nes to typescript
