+++ packages/core-p2p/src/socket-server/server.ts
@@ -1,6 +1,6 @@
 import { Container, Contracts, Types } from "@arkecosystem/core-kernel";
 import { Server as HapiServer, ServerInjectOptions, ServerInjectResponse, ServerRoute } from "@hapi/hapi";
-import Nes from "../hapi-nes";
+import { plugin } from "../hapi-nes";
 
 import { AcceptPeerPlugin } from "./plugins/accept-peer";
 import { ValidatePlugin } from "./plugins/validate";
@@ -43,7 +43,7 @@ export class Server {
         this.server = new HapiServer({ address: optionsServer.hostname, port: optionsServer.port });
         (this.server.app as any).app = this.app;
 
-        await this.server.register(Nes);
+        await this.server.register({ plugin });
 
         this.app.resolve(InternalRoute).register(this.server);
         this.app.resolve(PeerRoute).register(this.server);
